# BackEnd/tshirt-store-api/tshirt-api/src/likes/likes.controller.ts
import {
  Controller,
  Post,
  Delete,
  Param,
  ParseIntPipe,
  UseGuards,
  HttpCode,
  HttpStatus,
  Get,
  Query,
} from '@nestjs/common';
import { ApiBearerAuth, ApiTags, ApiOperation } from '@nestjs/swagger';
import { LikesService } from './likes.service';
import { JwtAuthGuard } from '../common/guards/jwt-auth.guard';
import { CurrentUser } from '../common/decorators/current-user.decorator';
import type { AuthenticatedUser } from '../common/types/authenticated-user.type';
import { ListLikedProductsQueryDto } from './dto/list-liked-products-query.dto';

@ApiTags('Likes')
@Controller('likes')
@UseGuards(JwtAuthGuard)
@ApiBearerAuth()
export class LikedProductsController {
  constructor(private likesService: LikesService) {}

  @Get('products')
  @ApiOperation({ summary: 'List products liked by the current user' })
  findLikedProducts(
    @CurrentUser() user: AuthenticatedUser,
    @Query() query: ListLikedProductsQueryDto,
  ) {
    return this.likesService.findLikedProducts(user.id, query);
  }
}

@ApiTags('Likes')
@Controller('products/:productId/like')
@UseGuards(JwtAuthGuard)
@ApiBearerAuth()
export class LikesController {
  constructor(private likesService: LikesService) {}

  @Post()
  @HttpCode(HttpStatus.CREATED)
  @ApiOperation({ summary: 'Like a product' })
  like(
    @Param('productId', ParseIntPipe) productId: number,
    @CurrentUser() user: AuthenticatedUser,
  ) {
    return this.likesService.like(user.id, productId);
  }

  @Delete()
  @HttpCode(HttpStatus.NO_CONTENT)
  @ApiOperation({ summary: 'Unlike a product' })
  unlike(
    @Param('productId', ParseIntPipe) productId: number,
    @CurrentUser() user: AuthenticatedUser,
  ) {
    return this.likesService.unlike(user.id, productId);
  }
}