# BackEnd/tshirt-store-api/week-2/openapi.yaml
openapi: 3.0.3
info:
  title: T-Shirt Store API
  description: |
    REST API for the T-Shirt Store capstone project.
    Supports three roles: Manager, Client, and Delivery Person.
  version: 1.0.0
  contact:
    name: Adrian
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ─── Reusable parameters ────────────────────────────────────────
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    ProductIdParam:
      name: productId
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
    OrderIdParam:
      name: orderId
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
  schemas:
...
    Category:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          example: Graphic Tees
        slug:
          type: string
          example: graphic-tees
        description:
          type: string

    # ── Address ──
...
    ProductSummary:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          example: Mountain Sunset Tee
        slug:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [active, disabled]
        category:
          $ref: "#/components/schemas/Category"
        primaryImage:
          type: string
          format: uri
          nullable: true
        likesCount:
          type: integer
        isLiked:
          type: boolean
          description: Whether the current user has liked this product (false if not logged in)
        minPrice:
          type: number
          format: float
          description: Lowest variant price for this product
        createdAt:
          type: string
          format: date-time
    ProductDetail:
      allOf:
        - $ref: "#/components/schemas/ProductSummary"
        - type: object
          properties:
            images:
              type: array
              items:
                $ref: "#/components/schemas/ProductImage"
            variants:
              type: array
              items:
                $ref: "#/components/schemas/ProductVariant"
    CreateProductRequest:
      type: object
      required: [name, description, categoryId]
      properties:
        name:
          type: string
        description:
          type: string
        categoryId:
          type: integer
          format: int64
...
    ProductImage:
      type: object
      properties:
        id:
          type: integer
          format: int64
        publicUrl:
          type: string
          format: uri
        altText:
          type: string
        sortOrder:
          type: integer
        isPrimary:
          type: boolean

    # ── Product Variant ──
    ProductVariant:
      type: object
      properties:
        id:
          type: integer
          format: int64
        productId:
          type: integer
          format: int64
        sizeId:
          type: integer
          format: int64
        colorId:
          type: integer
          format: int64
        sku:
          type: string
          example: MST-BLU-L
        price:
          type: number
          format: float
          example: 29.99
        stock:
          type: integer
          example: 50
        isActive:
          type: boolean
        size:
          type: object
          properties:
            id:
              type: integer
              format: int64
            name:
              type: string
              example: L
        color:
          type: object
          properties:
            id:
              type: integer
              format: int64
            name:
              type: string
              example: Blue
            hexCode:
              type: string
              nullable: true
              example: "#0000FF"
...
    OrderSummary:
      type: object
      properties:
        id:
          type: integer
          format: int64
        orderNumber:
          type: string
        orderStatus:
          type: string
          enum: [pending, paid, processing, shipped, delivered, cancelled]
        subtotal:
          type: number
          format: float
        discountAmount:
          type: number
          format: float
        totalAmount:
          type: number
          format: float
        paymentMethod:
          type: string
          enum: [payment_link, payment_intent]
          nullable: true
        createdAt:
          type: string
          format: date-time
    OrderDetail:
      allOf:
        - $ref: "#/components/schemas/OrderSummary"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/OrderItem"
            promoCode:
              type: string
              nullable: true
            promoDiscount:
              type: number
              format: float
              nullable: true
            shippingAddress:
              type: object
              properties:
                recipientName:
                  type: string
                recipientPhone:
                  type: string
                line1:
                  type: string
                line2:
                  type: string
                  nullable: true
                city:
                  type: string
                stateRegion:
                  type: string
                  nullable: true
                postalCode:
                  type: string
                countryCode:
                  type: string
            statusHistory:
              type: array
              items:
                $ref: "#/components/schemas/StatusHistoryEntry"
    OrderItem:
      type: object
      properties:
        productName:
          type: string
        skuCode:
          type: string
        sizeName:
          type: string
        colorName:
          type: string
        imageUrl:
          type: string
          format: uri
          nullable: true
        quantity:
          type: integer
        unitPrice:
          type: number
          format: float
        lineTotal:
          type: number
          format: float
    StatusHistoryEntry:
      type: object
      properties:
        orderStatus:
          type: string
          enum: [pending, paid, processing, shipped, delivered, cancelled]
        changedBy:
          type: string
          nullable: true
          description: User email or "system"
        reason:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
    CreateOrderRequest:
      type: object
      required: [addressId]
      properties:
        addressId:
          type: integer
          format: int64
        promoCode:
          type: string
          description: Optional promo code to apply
    UpdateOrderStatusRequest:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [processing, shipped, delivered, cancelled]
        reason:
          type: string

    # ── Payment ──
    CreatePaymentIntentResponse:
      type: object
      properties:
        clientSecret:
          type: string
          description: Stripe client secret for frontend confirmation
        paymentIntentId:
          type: string
        amount:
          type: number
          format: float
        currency:
          type: string
          example: usd
    CreatePaymentLinkRequest:
      type: object
      required: [productVariantId, quantity, addressId]
      properties:
        productVariantId:
          type: integer
          format: int64
        quantity:
          type: integer
          minimum: 1
        addressId:
          type: integer
          format: int64
    CreatePaymentLinkResponse:
      type: object
      properties:
        paymentLinkUrl:
          type: string
          format: uri
        orderId:
          type: integer
          format: int64
        demo:
          type: boolean
          description: Present only in local development when Stripe uses the placeholder key
...
    PaginatedResponse:
      type: object
      properties:
        data:
          type: array
          items: {}
        meta:
          type: object
          properties:
            page:
              type: integer
            limit:
              type: integer
            totalItems:
              type: integer
            totalPages:
              type: integer

    # ── Error ──
    ErrorResponse:
      type: object
      properties:
        statusCode:
          type: integer
        message:
          type: string
        error:
          type: string
      example:
        statusCode: 400
        message: Validation failed
        error: Bad Request

# ─── Paths ──────────────────────────────────────────────────────────
paths:
...
  /products:
    get:
      tags: [Products]
      summary: List products with pagination (public)
      operationId: listProducts
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
        - name: categoryId
          in: query
          schema:
            type: integer
            format: int64
          description: Filter by category
        - name: search
          in: query
          schema:
            type: string
          description: Search by product name
      responses:
        "200":
          description: Paginated product list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PaginatedResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/ProductSummary"

    post:
      tags: [Products]
      summary: Create a product (manager only)
      operationId: createProduct
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductRequest"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDetail"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not authorized (not a manager)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /products/{productId}/like:
    post:
      tags: [Likes]
      summary: Like a product (client only)
      operationId: likeProduct
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ProductIdParam"
      responses:
        "201":
          description: Product liked
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Already liked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      tags: [Likes]
      summary: Unlike a product (client only)
      operationId: unlikeProduct
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ProductIdParam"
      responses:
        "204":
          description: Like removed
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Like not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Cart ──────────────────────────────────────────────────────────
...
  /orders:
    post:
      tags: [Orders]
      summary: Create an order from cart (client only)
      operationId: createOrder
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "201":
          description: Order created (status pending)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderDetail"
        "400":
          description: Cart is empty or stock unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    get:
      tags: [Orders]
      summary: List orders (client sees own, manager sees all)
      operationId: listOrders
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, paid, processing, shipped, delivered, cancelled]
        - name: fromDate
          in: query
          schema:
            type: string
            format: date
          description: Filter orders from this date
        - name: toDate
          in: query
          schema:
            type: string
            format: date
          description: Filter orders up to this date
        - name: minAmount
          in: query
          schema:
            type: number
            format: float
          description: Minimum total amount
        - name: maxAmount
          in: query
          schema:
            type: number
            format: float
          description: Maximum total amount
      responses:
        "200":
          description: Paginated order list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PaginatedResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/OrderSummary"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/{orderId}:
    get:
      tags: [Orders]
      summary: Get order details (client own order, manager any order)
      operationId: getOrder
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/OrderIdParam"
      responses:
        "200":
          description: Order detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderDetail"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not authorized to view this order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/{orderId}/status:
    patch:
      tags: [Orders]
      summary: Update order status (manager or delivery person)
      operationId: updateOrderStatus
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/OrderIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusRequest"
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderDetail"
        "400":
          description: Invalid status transition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not authorized for this transition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/{orderId}/cancel:
    post:
      tags: [Orders]
      summary: Cancel an order (client, only before shipped)
      operationId: cancelOrder
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/OrderIdParam"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
      responses:
        "200":
          description: Order cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderDetail"
        "400":
          description: Order cannot be cancelled (already shipped or delivered)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not authorized (not the order owner)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Payments ──────────────────────────────────────────────────────
  /orders/{orderId}/payment-intent:
    post:
      tags: [Payments]
      summary: Create a Stripe Payment Intent for cart checkout (client only)
      operationId: createPaymentIntent
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/OrderIdParam"
      responses:
        "201":
          description: Payment Intent created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatePaymentIntentResponse"
        "400":
          description: Order not in pending status or stock unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not the order owner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/{orderId}/payment-link:
    post:
      tags: [Payments]
      summary: Create a Stripe Checkout link for an existing order (client only)
      operationId: createOrderPaymentLink
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/OrderIdParam"
      responses:
        "201":
          description: Payment Link created for the existing order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatePaymentLinkResponse"
        "400":
          description: Order is not pending
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not the order owner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /payments/payment-link:
    post:
      tags: [Payments]
      summary: Create a Stripe Payment Link for single product purchase (client only)
      operationId: createPaymentLink
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaymentLinkRequest"
      responses:
        "201":
          description: Payment Link created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatePaymentLinkResponse"
        "400":
          description: Insufficient stock
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Webhooks ──────────────────────────────────────────────────────
...